Reconnect logic extracted from PoloniexStreamingService to NettyStreamingService (#56)
* Fixed NPEs incorrect subscriptions for Bitmex, Bitfinex, Poloniex

* Reconnect logic extracted from PoloniexStreamingService to NettyStreamingService

* Merge branch 'develop' of https://github.com/bitrich-info/xchange-stream into develop

# Conflicts:
#	service-netty/src/main/java/info/bitrich/xchangestream/service/netty/NettyStreamingService.java
#	xchange-bitfinex/src/main/java/info/bitrich/xchangestream/bitfinex/BitfinexStreamingService.java
#	xchange-poloniex2/src/main/java/info/bitrich/xchangestream/poloniex2/PoloniexStreamingService.java

* Merge branch 'develop' of https://github.com/bitrich-info/xchange-stream into develop

# Conflicts:
#	service-netty/src/main/java/info/bitrich/xchangestream/service/netty/NettyStreamingService.java
#	xchange-bitfinex/src/main/java/info/bitrich/xchangestream/bitfinex/BitfinexStreamingService.java
#	xchange-poloniex2/src/main/java/info/bitrich/xchangestream/poloniex2/PoloniexStreamingService.java
