+++ xchange-poloniex2/src/main/java/info/bitrich/xchangestream/poloniex2/PoloniexStreamingService.java
@@ -7,9 +7,6 @@ import info.bitrich.xchangestream.poloniex2.dto.PoloniexWebSocketEvent;
 import info.bitrich.xchangestream.poloniex2.dto.PoloniexWebSocketEventsTransaction;
 import info.bitrich.xchangestream.poloniex2.dto.PoloniexWebSocketSubscriptionMessage;
 import info.bitrich.xchangestream.service.netty.JsonNettyStreamingService;
-import info.bitrich.xchangestream.service.netty.WebSocketClientHandler;
-import io.netty.channel.ChannelHandlerContext;
-import io.netty.handler.codec.http.websocketx.WebSocketClientHandshaker;
 import io.reactivex.Completable;
 import io.reactivex.Observable;
 import org.knowm.xchange.currency.CurrencyPair;
@@ -31,7 +28,7 @@ public class PoloniexStreamingService extends JsonNettyStreamingService {
 
     private final Map<String, String> subscribedChannels = new HashMap<>();
     private final Map<String, Observable<JsonNode>> subscriptions = new HashMap<>();
-    private boolean isManualDisconnect = false;
+
 
     public PoloniexStreamingService(String apiUrl) {
         super(apiUrl, Integer.MAX_VALUE);
@@ -134,33 +131,8 @@ public class PoloniexStreamingService extends JsonNettyStreamingService {
 
     @Override
     public Completable disconnect() {
-        isManualDisconnect = true;
+
         return super.disconnect();
     }
 
-    @Override
-    protected WebSocketClientHandler getWebSocketClientHandler(WebSocketClientHandshaker handshaker,
-                                                               WebSocketClientHandler.WebSocketMessageHandler handler) {
-        LOG.info("Registering Poloniex2WebSocketClientHandler");
-        return new Poloniex2WebSocketClientHandler(handshaker, handler);
-    }
-
-    private class Poloniex2WebSocketClientHandler extends WebSocketClientHandler {
-        Poloniex2WebSocketClientHandler(WebSocketClientHandshaker handshaker, WebSocketMessageHandler handler) {
-            super(handshaker, handler);
-        }
-
-        @Override
-        public void channelInactive(ChannelHandlerContext ctx) {
-            if (isManualDisconnect) {
-                isManualDisconnect = false;
-            } else {
-                super.channelInactive(ctx);
-                LOG.info("Reopening websocket because it was closed by the host");
-                connect().blockingAwait();
-                LOG.info("Resubscribing channels");
-                resubscribeChannels();
-            }
-        }
-    }
 }
